Not remove extension if other depends on this one

+++ src/ExtensionHandler.ts
@@ -67,17 +67,32 @@ export default class ExtensionHandler {
             if (!selected) return;
             let packageName = selected.path;
 
-            for (let file of fs.readdirSync(this.extensionFolder + packageName))
-                fs.unlinkSync(this.extensionFolder + packageName + "/" + file);
-            fs.rmdirSync(this.extensionFolder + packageName);
-            Logger.println("Removed extension: " + packageName);
-            
-            let f = packageName.split('/');
+            // Check if an installed extension depends on this one
+            let ableToRemove = true
+            extensionsData.forEach(element => {
+                if (element.dependencies) {
+                    element.dependencies.forEach(d => {
+                        if (d == packageName) {
+                            window.showErrorMessage(`Can't remove ${selected.name} extension, ${element.name} depends on this extension`);
+                            ableToRemove = false;
+                        }
+                    });
+                }
+            });
 
-            if (f.length > 2) {
-                let rootFolder = fs.readdirSync(this.extensionFolder + f[1]);
-                if (rootFolder.length == 0)
+            if (ableToRemove) {
+                for (let file of fs.readdirSync(this.extensionFolder + packageName))
+                fs.unlinkSync(this.extensionFolder + packageName + "/" + file);
+                fs.rmdirSync(this.extensionFolder + packageName);
+                Logger.println("Removed extension: " + packageName);
+                
+                let f = packageName.split('/');
+                
+                if (f.length > 2) {
+                    let rootFolder = fs.readdirSync(this.extensionFolder + f[1]);
+                    if (rootFolder.length == 0)
                     fs.rmdirSync(this.extensionFolder + f[1]);
+                }
             }
         });
